Refactor `ociartifact` to simplify artifact creation using `NewArtifact` and remove redundant `buildArtifact` logic.
(cherry picked from commit 39ff6f590a23485c1a8b5941bf4c2387bd6b9b64)
